Clean up OpenTAP loading implementation in tasks. Add API for deinitializing session logs.
