+++ sdk/OpenTap.Sdk.MSBuild/InstallOpenTapPackages.cs
@@ -118,31 +118,10 @@ namespace Keysight.OpenTap.Sdk.MSBuild
 
             Environment.SetEnvironmentVariable("OPENTAP_NO_UPDATE_CHECK", "true");
             Environment.SetEnvironmentVariable("OPENTAP_DEBUG_INSTALL", "true");
-
-            var thisAsmDir = Path.GetDirectoryName(typeof(OpenTapImageInstaller).Assembly.Location);
-            var openTapDll = Path.Combine(thisAsmDir, "payload", "OpenTap.dll");
-            var openTapPackageDll = Path.Combine(thisAsmDir, "payload", "OpenTap.Package.dll");
-
-            // This is sort of a hack because the standard resolver will try to resolve OpenTap 9.4.0.0,
-            // but we need to load whatever is in the NuGet directory
-            Assembly resolve(object sender, ResolveEventArgs args)
-            {
-                if (args.Name.StartsWith("OpenTap.Package"))
-                    return Assembly.LoadFile(openTapPackageDll);
-                if (args.Name.StartsWith("OpenTap"))
-                    return Assembly.LoadFile(openTapDll);
-                return null;
-            }
-
-            try
-            {
-                AppDomain.CurrentDomain.AssemblyResolve += resolve;
+            
+            // Ensure the correct OpenTAP dll is loaded, and the session log is disposed properly.
+            using (OpenTapContext.Create())
                 return InstallPackages();
-            }
-            finally
-            {
-                AppDomain.CurrentDomain.AssemblyResolve -= resolve;
-            }
         }
 
         private CancellationTokenSource cts = new CancellationTokenSource();
